requirements.txt
streamlit
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0
shap>=0.44.0
xgboost>=2.0.0
lightgbm>=4.0.0
optuna>=3.5.0
streamlit-option-menu>=0.3.6
Pillow>=10.0.0
wandb>=0.17.0
python-dotenv>=1.0.0